+++ src/index.js
@@ -1,5 +1,5 @@
 function now() {
-  return window.performance ? window.performance.now() : Date.now ? Date.now() : new Date().getTime();
+  return window && window.performance ? window.performance.now() : Date.now;
 }
 
 function tick(instance) {
@@ -20,24 +20,40 @@ class Timer {
    * Timer constructor
    *
    * @constructor Timer
-   * @param {Number} [duration] The timer's duration (ms). If left `undefined` or `0` or negative number the timer counts up instead of down.
-   * @param {Function} [callback] Function to be executed while timer is running. The Timer instance is passed by as parameter.
+   * @param {Number} elapsedTime The time that has elapsed in milliseconds. If negative number provided, it will become `0`. If a number greater than `duration` is provided, it will become equal to `duration`.
+   * @param {Number} duration The timer's duration in milliseconds. If negative number provided, it will become `0`.
+   * @param {Function} [callback] Function to be executed while timer is running. The `Timer` instance is passed by as parameter.
+   * @throws {TypeError} If `duration` is not a number or `NaN`.
+   * @throws {TypeError} If `elapsedTime` is not a number or `NaN`.
    */
-  constructor(duration, callback) {
+  constructor(elapsedTime, duration, callback) {
+    if (typeof elapsedTime !== 'number' || Number.isNaN(elapsedTime)) {
+      throw new TypeError('Expected a number for "elapsedTime"');
+    }
+
+    if (typeof duration !== 'number' || Number.isNaN(duration)) {
+      throw new TypeError('Expected a number for "duration"');
+    }
+
     this._started = false;
     this._now = 0;
-    this._time = 0;
+    this._elapsedTime = elapsedTime;
     this._duration = duration;
     this._callback = callback;
 
-    if (duration < 0) {
+    if (this._duration < 0) {
       this._duration = 0;
     }
 
-    if (!duration || typeof duration === 'function') {
-      this._duration = 0;
-      this._callback = duration;
+    if (this._elapsedTime > this._duration) {
+      this._elapsedTime = this._duration;
     }
+
+    if (this._elapsedTime < 0) {
+      this._elapsedTime = 0;
+    }
+
+    this._time = this._elapsedTime;
   }
 
   /**
@@ -69,7 +85,7 @@ class Timer {
       this.reset(true);
     }
 
-    if (this._started || Number(this._duration) && this._time > this._duration) {
+    if (this._started || this._duration >= 0 && this._time >= this._duration) {
       return this;
     }
 
@@ -104,7 +120,7 @@ class Timer {
       this.stop();
     }
 
-    this._time = 0;
+    this._time = this._elapsedTime;
     return this;
   }