+++ test/Timer.spec.js
@@ -3,35 +3,23 @@ import Timer from '../src/index';
 let timer, elapsed, remaining;
 const onTimerRunning = instance => ({ elapsed, remaining } = instance.time());
 
-beforeEach(() => {
-  elapsed = remaining = 0;
-  timer && timer.stop();
-  timer = new Timer(200, onTimerRunning);
-});
+describe('Timer (normal flow)', () => {
+  beforeEach(() => {
+    elapsed = remaining = 0;
+    timer && timer.stop();
+    timer = new Timer(100, 200, onTimerRunning);
+  });
 
-describe('Timer', () => {
   it('initializes a timer', () => {
     expect(timer instanceof Timer).toBe(true);
   });
 
-  it('initializes a timer with no duration or using a negative number', done => {
-    const t1 = new Timer(onTimerRunning);
-    const t2 = new Timer(-1000, onTimerRunning);
+  it('test elapsedTime', () => {
+    expect(timer._elapsedTime).toBe(100);
+  });
 
-    t1.start();
-    t2.start();
-
-    setTimeout(() => {
-      expect(t1.time().remaining).toBe(0);
-      expect(t1.time().elapsed).toBeGreaterThan(0);
-
-      expect(t2.time().remaining).toBe(0);
-      expect(t2.time().elapsed).toBeGreaterThan(0);
-
-      t1.stop();
-      t2.stop();
-      done();
-    }, 200);
+  it('test duration', () => {
+    expect(timer._duration).toBe(200);
   });
 
   it('starts the timer without reset', () => {
@@ -40,8 +28,11 @@ describe('Timer', () => {
   });
 
   it('starts the timer and then resets', done => {
-    timer.start(true);
-    expect(timer._started).toBe(true);
+    const t = new Timer(0, 200);
+
+    t.start(true);
+
+    expect(t._started).toBe(true);
 
     setTimeout(() => {
       expect(timer._started).toBe(false);
@@ -101,5 +92,49 @@ describe('Timer', () => {
       expect(timer.isRunning()).toBe(false);
       done();
     }, 500);
+  });
+});
+
+describe('Timer (edge cases)', () => {
+  it ('throws if elapsedTime is not a number', () => {
+    expect(() => {
+      return new Timer('0', 200);
+    }).toThrow(new TypeError('Expected a number for "elapsedTime"'));
+  });
+
+  it ('throws if elapsedTime is NaN', () => {
+    expect(() => {
+      return new Timer(NaN, 200);
+    }).toThrow(new TypeError('Expected a number for "elapsedTime"'));
+  });
+
+  it ('throws if duration is not a number', () => {
+    expect(() => {
+      return new Timer(0, '200');
+    }).toThrow(new TypeError('Expected a number for "duration"'));
+  });
+
+  it ('throws if duration is NaN', () => {
+    expect(() => {
+      return new Timer(0, NaN);
+    }).toThrow(new TypeError('Expected a number for "duration"'));
+  });
+
+  it('duration becomes 0 if negative', () => {
+    const t = new Timer(0, -1000);
+
+    expect(t._duration).toBe(0);
+  });
+
+  it('elapsedTime becomes 0 if negative', () => {
+    const t = new Timer(-1000, 1000);
+
+    expect(t._elapsedTime).toBe(0);
+  });
+
+  it('elapsedTime is greater than duration', () => {
+    const t = new Timer(1000, 0);
+
+    expect(t._elapsedTime).toBe(0);
   });
 });